[refactor] components/ 내부 폴더링 수정
