+++ src/components/common/Header.tsx
@@ -4,10 +4,10 @@ import HeaderLogo from '../assets/header_logo.svg';
 import InvitationIcon from '../assets/icon/invitation.svg';
 import BuddyIcon from '../assets/icon/buddy.svg';
 import MenuIcon from '../assets/icon/menu.svg';
-import SideModal from './modal/SideModal';
-import MypageContent from './modal/MyPageContent';
-import BuddyContent from './modal/BuddyContent';
-import InvitationContent from './modal/InvitationContent';
+import SideModal from '../modal/SideModal';
+import MypageContent from '../modal/MyPageContent';
+import BuddyContent from '../modal/BuddyContent';
+import InvitationContent from '../modal/InvitationContent';
 
 const Header = () => {
   const navigate = useNavigate();
