+++ src/components/pages/LoginPage.tsx
@@ -1,10 +1,10 @@
 import { GoogleOAuthProvider, TokenResponse, useGoogleLogin } from '@react-oauth/google';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import BackGroundImage from './BackGroundImage'
+import BackGroundImage from '../common/BackGroundImage'
 import Logo from '../assets/logo.svg';
 import LoginButton from '../assets/button/login_button.svg'
-import IsLoggedIn from './IsLoggedIn';
+import IsLoggedIn from '../utils/IsLoggedIn';
 
 const LoginPage = () => {
 	const navigate = useNavigate()
